refactor: formatRelativeTime 유틸을 공통 유틸로 분리
