+++ src/features/board/utils/index.ts
@@ -1,4 +1,4 @@
-export { formatRelativeTime } from "./formatRelativeTime";
+export { formatRelativeTime } from "@/utils";
 export { mapFormToCreatePostRequest } from "./mapFormToCreatePostRequest";
 export { mapPostDetailToFormValues } from "./mapPostDetailToFormValues";
 export { mapSharedLifestyleToHabits } from "./mapSharedLifestyleToHabits";
